custom_users/project/requirements.txt
anyascii==0.3.2
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
defusedxml==0.7.1
Django==5.0.3
django-filter==24.2
django-modelcluster==6.3
django-permissionedforms==0.1
django-taggit==5.0.1
django-treebeard==4.7.1
djangorestframework==3.15.2
draftjs-exporter==5.0.0
env==0.1.0
et-xmlfile==1.1.0
filetype==1.2.0
idna==3.7
l18n==2021.3
laces==0.1.1
openpyxl==3.1.5
pillow==10.3.0
pillow_heif==0.16.0
pytz==2024.1
requests==2.32.3
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
telepath==0.3.1
tzdata==2024.1
urllib3==2.2.2
wagtail==6.1.2
Willow==1.8.0
